+++ arrow-libs/core/arrow-core-retrofit/build.gradle
@@ -13,9 +13,4 @@ dependencies {
     compileOnly "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$KOTLIN_VERSION"
     compileOnly project(":arrow-core")
     compileOnly "com.squareup.retrofit2:retrofit:$RETROFIT_VERSION"
-    testCompileOnly "org.jetbrains.kotlin:kotlin-reflect:$KOTLIN_VERSION"
-    testRuntimeOnly "org.junit.vintage:junit-vintage-engine:$JUNIT_VINTAGE_VERSION"
-    testImplementation project(":arrow-core-test")
-    testImplementation "com.squareup.retrofit2:converter-gson:$RETROFIT_VERSION"
-    testImplementation "com.squareup.okhttp3:mockwebserver:$MOCKWEBSERVER_VERSION"
 }
