+++ arrow-libs/core/arrow-core-test/src/commonTest/kotlin/arrow/continuations/SuspendingComputationTest.kt
@@ -1,8 +1,10 @@
-package generic
+package arrow.continuations
 
 import arrow.core.computations.either
 import arrow.core.Either.Right
 import arrow.core.Either.Left
+import arrow.core.Some
+import arrow.core.computations.option
 import io.kotest.assertions.fail
 import io.kotest.matchers.shouldBe
 import io.kotest.assertions.throwables.shouldThrow
@@ -71,7 +73,7 @@ class SuspendingComputationTest : StringSpec({
 
   "Can short-circuit immediately from nested blocks" {
     either<String, Int> {
-      val x = maybeEff {
+      val x = option {
         Left("test").bind()
         5L
       }
@@ -83,7 +85,7 @@ class SuspendingComputationTest : StringSpec({
 
   "Can short-circuit suspended from nested blocks" {
     either<String, Int> {
-      val x = maybeEff {
+      val x = option {
         Left("test").suspend().bind()
         5L
       }
@@ -95,8 +97,8 @@ class SuspendingComputationTest : StringSpec({
 
   "Can short-circuit immediately after suspending from nested blocks" {
     either<String, Int> {
-      val x = maybeEff {
-        Just(1L).suspend().bind()
+      val x = option {
+        Some(1L).suspend().bind()
         Left("test").suspend().bind()
         5L
       }
@@ -108,8 +110,8 @@ class SuspendingComputationTest : StringSpec({
 
   "Can short-circuit suspended after suspending from nested blocks" {
     either<String, Int> {
-      val x = maybeEff {
-        Just(1L).suspend().bind()
+      val x = option {
+        Some(1L).suspend().bind()
         Left("test").suspend().bind()
         5L
       }
