+++ arrow-libs/core/arrow-core-test/src/commonTest/kotlin/arrow/core/FunctionSyntaxTest.kt
