+++ arrow-libs/core/arrow-core-test/src/jvmTest/java/arrow/core/ValidatedUsage.java
