+++ arrow-libs/core/arrow-core-test/src/jvmTest/java/arrow/retrofit/adapter/mock/ErrorMock.kt
