+++ arrow-libs/core/arrow-core-test/src/jvmTest/java/arrow/typeclasses/SemigroupUsageTest.java
